small fixes, new slides 5-6

+++ CourseSessions/Sessions45/R/runcode.R
@@ -18,16 +18,19 @@ file.copy( htmloutput, paste(local_directory,dochtmloutput, sep="/"), overwrite 
 setwd( "../" )
 unlink( "TMPdirReport", recursive = TRUE )      
 
-reportfilename = paste(slides_file, "Rmd", sep=".")
-docreportfilename = paste("doc", reportfilename, sep="/")
-htmloutput = paste(slides_file, "html", sep = ".")
-dochtmloutput = paste("doc", htmloutput, sep="/")
-
-unlink( "TMPdirSlides", recursive = TRUE )      
-dir.create( "TMPdirSlides" )
-setwd( "TMPdirSlides" )
-file.copy( paste(local_directory,docreportfilename, sep="/"),reportfilename, overwrite = T )
-slidify( reportfilename )
-file.copy( htmloutput, paste(local_directory,dochtmloutput, sep="/"), overwrite = T )
-setwd( "../" )
-unlink( "TMPdirSlides", recursive = TRUE )      
+if (0){
+  reportfilename = paste(slides_file, "Rmd", sep=".")
+  docreportfilename = paste("doc", reportfilename, sep="/")
+  htmloutput = paste(slides_file, "html", sep = ".")
+  dochtmloutput = paste("doc", htmloutput, sep="/")
+  
+  unlink( "TMPdirSlides", recursive = TRUE )      
+  dir.create( "TMPdirSlides" )
+  setwd( "TMPdirSlides" )
+  file.copy( paste(local_directory,docreportfilename, sep="/"),reportfilename, overwrite = T )
+  slidify( reportfilename )
+  file.copy( htmloutput, paste(local_directory,dochtmloutput, sep="/"), overwrite = T )
+  setwd( "../" )
+  unlink( "TMPdirSlides", recursive = TRUE )     
+  
+}
